fix: disable the native armcb isp driver
There are two ISP drivers provided by CIX. One is a native driver in the
kernel, called armcb_isp.ko when compiled as a module. The other one is
called armcb_isp_v4l2.ko, which is maintained as an external module[0].
The armcb_isp driver does not support V4L2 applications, such as Cheese
or GStreamer, while the armcb_isp_v4l2 driver does.

The armcb_isp was compiled as modules by default[1]. When both armcb_isp
and armcb_isp_v4l2 drivers exist on the system, armcb_isp was found to
be preferred, so the V4L2 applications don't work by default.

To solve the issue, we disable the armcb_isp driver in the kernel
config.

[0]: https://gitlab.com/cix-linux/cix_opensource/isp_driver/
[1]: https://github.com/radxa/kernel/blob/7a822ac9aea389869a28f259b507ceaf5fa4f527/arch/arm64/configs/cix.config#L180-L181

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>